refactor part of rti_addr handling code out of reactor_common
