+++ core/federated/federate.c
@@ -47,6 +47,7 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "../platform.h"
 #include "clock-sync.c" // Defines clock synchronization functions.
 #include "federate.h"   // Defines federate_instance_t
+#include "regex.h"
 
 // Error messages.
 char* ERROR_SENDING_HEADER = "ERROR sending header information to federate via RTI";
@@ -82,6 +83,15 @@ federate_instance_t _fed = {
         .triggers_for_network_input_control_reactions_size = 0,
         .trigger_for_network_output_control_reactions = NULL
 };
+
+
+federation_metadata_t federation_metadata = {
+    .federation_id =  "Unidentified Federation",
+    .rti_host = NULL,
+    .rti_port = -1,
+    .rti_user = NULL
+};
+
 
 /** 
  * Thread that listens for inputs from other federates.
@@ -2640,4 +2650,141 @@ tag_t _lf_send_next_event_tag(tag_t tag, bool wait_for_reply) {
         // TAN should be sent or a NET.
         tag = get_next_event_tag();
     }
+}
+
+
+
+/**
+ * Checks if port is valid.
+ * @return true if valid, false otherwise.
+ */
+bool validate_port(char* port) {
+    // magic number 6 since port range is [0, 65535]
+    int port_len = strnlen(port, 6); 
+    if (port_len < 1 || port_len > 5) {
+        return false;
+    }
+
+    for (int i = 0; i < port_len; i++) {
+        if (!isdigit(port[i])) {
+            return false;
+        }
+    }
+    int port_number = atoi(port);
+    return port_number >= 0 && port_number <= 65535;
+}
+
+/**
+ * Extract match groups from the rti_addr regex  
+ */
+bool extract_match_group(char* rti_addr, char* dest, regmatch_t group, int max_len, int min_len, char* err_msg) {
+    size_t size = group.rm_eo - group.rm_so;
+    if (size > max_len || size < min_len) {
+        error_print(err_msg);
+        return false;
+    }
+    strncpy(dest, &rti_addr[group.rm_so], size);
+    dest[size] = '\0';
+    return true;
+}
+
+/**
+ * Extract the host, port and user from rti_addr.  
+ */
+void extract_rti_addr_info(char* rti_addr, rti_addr_info_t* rti_addr_info) {
+    char* regex_str = "(([a-zA-Z0-9_-]{1,254})@)?([a-zA-Z0-9.]{1,255})(:([0-9]{1,5}))?";
+    size_t max_groups = 6;
+    // The group indices of each field of interest in the regex.
+    int user_gid = 2, host_gid = 3, port_gid = 5;
+    int gids[3] = {user_gid, host_gid, port_gid};
+
+    regex_t regex_compiled;
+    regmatch_t group_array[max_groups];
+
+    if (regcomp(&regex_compiled, regex_str, REG_EXTENDED)) {
+        error_print("Could not compile regex to parse RTI address");
+        return;
+    }
+
+    if (regexec(&regex_compiled, rti_addr, max_groups, group_array, 0) == 0) {
+        // Check for matched username. group_array[0] is the entire matched string.
+        for (int i = 1; i < max_groups; i++) {
+            DEBUG_PRINT("runtime rti_addr regex: so: %d   eo: %d\n", group_array[i].rm_so, group_array[i].rm_eo);
+        }
+        if (group_array[user_gid].rm_so != -1) {
+            if (!extract_match_group(rti_addr, rti_addr_info->rti_user_str, group_array[user_gid], 255, 1, "User name must be between 1 to 255 characters long.")) {
+                memset(rti_addr_info, 0, sizeof(rti_addr_info_t));
+                regfree(&regex_compiled);
+                return;
+            } else {
+                rti_addr_info->has_user = true;
+                DEBUG_PRINT("runtime rti_addr user: '%s'\n", rti_addr_info->rti_user_str);
+            }
+        }
+
+        // Check for matched host
+        if (group_array[host_gid].rm_so != -1) {
+            if (!extract_match_group(rti_addr, rti_addr_info->rti_host_str, group_array[host_gid], 255, 1, "Host must be between 1 to 255 characters long.")) {
+                memset(rti_addr_info, 0, sizeof(rti_addr_info_t));
+                regfree(&regex_compiled);
+                return;
+            } else {
+                rti_addr_info->has_host = true;
+                DEBUG_PRINT("runtime rti_addr host: '%s'\n", rti_addr_info->rti_host_str);
+            }
+        }
+
+        // Check for matched port
+        if (group_array[port_gid].rm_so != -1) {
+            if (!extract_match_group(rti_addr, rti_addr_info->rti_port_str, group_array[port_gid], 5, 1, "Port must be between 1 to 5 characters long.")) {
+                memset(rti_addr_info, 0, sizeof(rti_addr_info_t));
+                regfree(&regex_compiled);
+                return;
+            } else {
+                rti_addr_info->has_port = true;
+                DEBUG_PRINT("runtime rti_addr port: '%s'\n", rti_addr_info->rti_port_str);
+            }
+        }
+    }
+    regfree(&regex_compiled);
+}
+
+/**
+ * Parse the address of the RTI and store them into the global federation_metadata struct.
+ * @return a parse_rti_code_t indicating the result of the parse.
+ */
+parse_rti_code_t parse_rti_addr(char* rti_addr) {
+    bool has_host = false, has_port = false, has_user = false;
+    rti_addr_info_t rti_addr_info = {0};
+    extract_rti_addr_info(rti_addr, &rti_addr_info);
+    if (!rti_addr_info.has_host && !rti_addr_info.has_port && !rti_addr_info.has_user) {
+        return FAILED_TO_PARSE;
+    }
+
+    if (rti_addr_info.has_host) {
+        char* rti_host = (char*) calloc(256, sizeof(char));
+        strncpy(rti_host, rti_addr_info.rti_host_str, 255);
+        federation_metadata.rti_host = rti_host;
+    }
+    if (rti_addr_info.has_port) {
+        if (validate_port(rti_addr_info.rti_port_str)) {
+            federation_metadata.rti_port = atoi(rti_addr_info.rti_port_str);
+            DEBUG_PRINT("rti_port: %d\n", federation_metadata.rti_port);
+        } else {
+            return INVALID_PORT;
+        }
+    }
+    if (rti_addr_info.has_user) {
+        char* rti_user = (char*) calloc(256, sizeof(char));
+        strncpy(rti_user, rti_addr_info.rti_user_str, 255);
+        federation_metadata.rti_user = rti_user;
+    }
+    return SUCCESS;
+}
+
+/**
+ * Sets the federation_id of this federate to fid.
+ */
+void set_federation_id(char* fid) {
+    federation_metadata.federation_id = fid;
 }
